mysite/requirements.txt
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Pillow==9.4.0
python-dotenv==0.21.1
Telethon==1.27.0
